requirements.txt
google-cloud-bigquery
google-cloud-secret-manager
requests
Flask
gunicorn
